third_party/WebKit/LayoutTests/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-underline-position-expected.txt

Test to make sure text-underline-position property returns values properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value:
PASS e.style.getPropertyCSSValue('text-underline-position') is null
PASS computedStyle.textUnderlinePosition is 'auto'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'auto'

Value '':
PASS e.style.getPropertyCSSValue('text-underline-position') is null
PASS computedStyle.textUnderlinePosition is 'auto'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'auto'

Initial value (explicit):
PASS e.style.textUnderlinePosition is 'initial'
PASS e.style.getPropertyCSSValue('text-underline-position').toString() is '[object CSSValue]'
PASS e.style.getPropertyCSSValue('text-underline-position').cssText is 'initial'
PASS computedStyle.textUnderlinePosition is 'auto'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'auto'

Value 'auto':
PASS e.style.textUnderlinePosition is 'auto'
PASS e.style.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS e.style.getPropertyCSSValue('text-underline-position').cssText is 'auto'
PASS computedStyle.textUnderlinePosition is 'auto'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'auto'

Value 'alphabetic':
PASS e.style.textUnderlinePosition is 'alphabetic'
PASS e.style.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS e.style.getPropertyCSSValue('text-underline-position').cssText is 'alphabetic'
PASS computedStyle.textUnderlinePosition is 'alphabetic'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'alphabetic'

Value 'under':
PASS e.style.textUnderlinePosition is 'under'
PASS e.style.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS e.style.getPropertyCSSValue('text-underline-position').cssText is 'under'
PASS computedStyle.textUnderlinePosition is 'under'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'under'

Ancestor inherits values from parent:
PASS e.style.getPropertyCSSValue('text-underline-position') is null
PASS computedStyle.textUnderlinePosition is 'under'
PASS computedStyle.getPropertyCSSValue('text-underline-position').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-underline-position').cssText is 'under'

Value 'auto alphabetic':
PASS e.style.getPropertyCSSValue('text-underline-position') is null

Value 'auto under':
PASS e.style.getPropertyCSSValue('text-underline-position') is null

Value 'under under':
PASS e.style.getPropertyCSSValue('text-underline-position') is null

Value 'under under under':
PASS e.style.getPropertyCSSValue('text-underline-position') is null

PASS successfullyParsed is true

TEST COMPLETE
